Longs.

+++ Smoke/app/src/main/java/org/purple/smoke/TcpNeighbor.java
@@ -305,16 +305,16 @@ public class TcpNeighbor extends Neighbor
 			m_error = true;
 		    }
 
-		    long bytesRead = i;
+		    long bytesRead = (long) i;
 
-		    if(bytesRead < 0 || m_error)
+		    if(bytesRead < 0L || m_error)
 		    {
 			m_error = true;
 			setError("A socket read() error occurred.");
 			disconnect();
 			return;
 		    }
-		    else if(bytesRead == 0)
+		    else if(bytesRead == 0L)
 			return;
 
 		    m_bytesRead.getAndAdd(bytesRead);
@@ -339,6 +339,6 @@ public class TcpNeighbor extends Neighbor
 		{
 		}
 	    }
-	}, 0, READ_SOCKET_INTERVAL, TimeUnit.MILLISECONDS);
+	}, 0L, READ_SOCKET_INTERVAL, TimeUnit.MILLISECONDS);
     }
 }
